Now he can load the enemy by the id
When you Create a Enemy, you need a position and ID.
This id is a number that represents the type of enemy that you want
load.

Exemple:
enemy_1-> id = 1 has: hp=10 and attack=20.
enemy_4-> id = 4 has: hp=40 and attack=70.

This code Load the enemy 4: enemy_4 =
App->entity_elements->CreateEnemy(iPoint(200, 400), 4);